big changes to structure of keymaps. Moved most settings to user_settings.c in each keymap, and moved the color script to origin.
