+++ keyboards/posey_split_v3/keymaps/lambo/keymap.c
@@ -1,5 +1,4 @@
 #include QMK_KEYBOARD_H
-#include "keymap_german.h"
 
 //Layers
 #define _QWERTY 0
@@ -8,96 +7,15 @@
 // //Custom Layers
 #define _TEMPRTY 1
 #define _SC2 6
+//Colors
+#define C_QWERTY HSV_WHITE
+#define C_SYMB HSV_GREEN
+#define C_NAV HSV_RED
+// //Custom Layers
+#define C_TEMPRTY HSV_BLUE
+#define C_SC2 HSV_YELLOW
 
-// Custom Macros
-#define TEMP_TYPE_LAYER _TEMPRTY
-#define REPEAT_ENABLED
-#define REPEAT_ALL_KEYS_ENABLED
-#define BOOSTED_REPEAT_ENABLED
-#define REPEAT_DELAY 108
-#define REPEAT_TERM 5
-#define BOOSTED_REPEAT_TERM 8
-#define BOOSTED_REPEAT_KEYS KC_Z
-#define BOOSTED_REPEAT_KEY_COUNT 1
-#define FAST_REPEAT_LAYERS _SC2
-#define FAST_REPEAT_LAYER_COUNT 1
-
-enum custom_keycodes {
-   QWERTY = SAFE_RANGE,
-   SYMB,
-   NAV,
-   ADJUST,
-   SC2,
-   //quick type
-   CTT,
-   TTC
-};
-
-#ifdef REPEAT_ENABLED
-   #ifndef REPEAT_DELAY
-      #define REPEAT_DELAY 100
-   #endif
-   #ifndef REPEAT_TERM
-      #define REPEAT_TERM 5
-   #endif
-   #include "hardware_repeat.c"
-#endif
-
-#ifdef QUICK_GAME_TYPE_ENABLED
-   #include "quick_game_type.c"
-#endif
-#ifndef QUICK_GAME_TYPE_ENABLED
-   #define TTC KC_ENTER
-   #define CTT KC_ENTER
-#endif
-
-#ifdef DIABLO_PLAYERS_CHANGE_ENABLED
-   #include "diablo_player_change.c"
-#endif
-
-// Shortcut to make keymap more readable
-#define SYM_L   MO(_SYMB)
-
-#define KC_ALAS LALT_T(KC_PAST)
-#define KC_CTPL LCTL_T(KC_BSLS)
-
-#define KC_NAGR LT(_NAV, KC_GRV)
-#define KC_NAMI LT(_NAV, KC_MINS)
-#define SYM_NAV LT(_NAV, KC_CAPS)
-#define KC_ADEN LT(_ADJUST, KC_END)
-#define KC_ADPU LT(_ADJUST, KC_PGUP)
-
-// Layer Switch Shortcuts
-#define SYM_SC2 TO(_SC2)
-#define SYM_Q TO(_QWERTY)
-#define SYM_DIA TO(_DIABLO)
-#define PL_C OSL(_PLAYERS)
-
-//German Key Alterations
-// #define LAMBO_1 DE_LABK
-// #define LAMBO_2 S(LAMBO_1)
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-   #ifdef DIABLO_PLAYERS_CHANGE_ENABLED
-   if (!process_diablo_player_change(keycode, record)) {
-      return false;
-   }
-   #endif
-
-   #ifdef QUICK_GAME_TYPE_ENABLED
-   if (!process_quick_game_type(keycode, record)) {
-      return false;
-   }
-   #endif
-
-   #ifdef REPEAT_ENABLED
-   if (!process_repeat_key(keycode, record)) {
-      return false;
-   }
-   #endif
-
-   return true;
-};
+#include "user_settings.c"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     /*
