+++ keyboards/posey_split_v3/keymaps/mrllama/keymap.c
@@ -11,104 +11,7 @@
 #define _DIABLO 8
 #define _PLAYERS 9
 
-// Custom Macros
-#define QUICK_GAME_TYPE_ENABLED
-#define TEMP_TYPE_LAYER _TEMPRTY
-#define DIABLO_PLAYERS_CHANGE_ENABLED
-#define REPEAT_ENABLED
-#define REPEAT_DELAY 150
-#define REPEAT_TERM 5
-#define FAST_REPEAT_KEYS    KC_SPC, \
-                            KC_A
-#define FAST_REPEAT_KEY_COUNT 2
-
-#define FAST_REPEAT_LAYERS _SC2
-#define FAST_REPEAT_LAYER_COUNT 1
-
-enum custom_keycodes {
-   QWERTY = SAFE_RANGE,
-   SYMB,
-   NAV,
-   ADJUST,
-   SC2,
-   THECORE,
-   DIABLO,
-   //quick type
-   CTT,
-   TTC,
-   //players
-   PL_C,
-   PL_1,
-   PL_2,
-   PL_3,
-   PL_4,
-   PL_5,
-   PL_6,
-   PL_7,
-   PL_8
-};
-
-#ifdef REPEAT_ENABLED
-   #ifndef REPEAT_DELAY
-      #define REPEAT_DELAY 100
-   #endif
-   #ifndef REPEAT_TERM
-      #define REPEAT_TERM 5
-   #endif
-   #include "hardware_repeat.c"
-#endif
-
-#ifdef QUICK_GAME_TYPE_ENABLED
-   #include "quick_game_type.c"
-#endif
-#ifndef QUICK_GAME_TYPE_ENABLED
-   #define TTC KC_ENTER
-   #define CTT KC_ENTER
-#endif
-
-#ifdef DIABLO_PLAYERS_CHANGE_ENABLED
-   #include "diablo_player_change.c"
-#endif
-
-// Shortcut to make keymap more readable
-#define SYM_L   MO(_SYMB)
-
-#define KC_ALAS LALT_T(KC_PAST)
-#define KC_CTPL LCTL_T(KC_BSLS)
-
-#define KC_NAGR LT(_NAV, KC_GRV)
-#define KC_NAMI LT(_NAV, KC_MINS)
-#define SYM_NAV LT(_NAV, KC_CAPS)
-#define KC_ADEN LT(_ADJUST, KC_END)
-#define KC_ADPU LT(_ADJUST, KC_PGUP)
-
-// Layer Switch Shortcuts
-#define SYM_SC2 TO(_SC2)
-#define SYM_Q TO(_QWERTY)
-#define SYM_DIA TO(_DIABLO)
-#define PL_C OSL(_PLAYERS)
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-   #ifdef QUICK_GAME_TYPE_ENABLED
-   if (!process_diablo_player_change(keycode, record)) {
-      return false;
-   }
-   #endif
-
-   #ifdef DIABLO_PLAYERS_CHANGE_ENABLED
-   if (!process_quick_game_type(keycode, record)) {
-      return false;
-   }
-   #endif
-
-   #ifdef REPEAT_ENABLED
-   if (!process_repeat_key(keycode, record)) {
-      return false;
-   }
-   #endif
-
-   return true;
-};
+#include "user_settings.c"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     /*
